Change the Dashboard page UI and reposition the welcome line
